Combined leetcodeUsername methods with User Repo and SQL Repo
