+++ src/main/java/com/patina/codebloom/common/db/repos/user/UserSqlRepository.java
@@ -88,6 +88,39 @@ public class UserSqlRepository implements UserRepository {
     }
 
     @Override
+    public User getUserByLeetcodeUsername(final String inputLeetcodeUsername) {
+        User user = null;
+        String sql = """
+                            SELECT id, "discordId", "discordName", "leetcodeUsername", "nickname", admin, "profileUrl"
+                            FROM "User"
+                            WHERE "leetcodeUsername" = ?
+                        """;
+
+        try (PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, inputLeetcodeUsername);
+            try (ResultSet rs = stmt.executeQuery()) {
+                if (rs.next()) {
+                    var id = rs.getString("id");
+                    var discordId = rs.getString("discordId");
+                    var discordName = rs.getString("discordName");
+                    var leetcodeUsername = rs.getString("leetcodeUsername");
+                    var nickname = rs.getString("nickname");
+                    var admin = rs.getBoolean("admin");
+                    var profileUrl = rs.getString("profileUrl");
+                    var tags = userTagRepository.findTagsByUserId(id);
+
+                    user = new User(id, discordId, discordName, leetcodeUsername, nickname, admin, profileUrl, tags);
+                    return user;
+                }
+            }
+        } catch (SQLException e) {
+            throw new RuntimeException("Error while retrieving user", e);
+        }
+
+        return user;
+    }
+
+    @Override
     public User getUserByDiscordId(final String inputDiscordId) {
         User user = null;
         String sql = "SELECT id, \"discordId\", \"discordName\", \"leetcodeUsername\", \"nickname\",\"schoolEmail\", admin, \"profileUrl\" FROM \"User\" WHERE \"discordId\"=?";
@@ -343,6 +376,54 @@ public class UserSqlRepository implements UserRepository {
             }
         } catch (SQLException e) {
             throw new RuntimeException("Failed to get user with score by id", e);
+        }
+
+        return null;
+    }
+
+    @Override
+    public UserWithScore getUserWithScoreByLeetcodeUsername(final String userLeetcodeUsername, final String leaderboardId) {
+        String sql = """
+                            SELECT
+                                u.id,
+                                u."discordId",
+                                u."discordName",
+                                u."leetcodeUsername",
+                                u.nickname,
+                                u.admin,
+                                u."profileUrl",
+                                m."totalScore"
+                            FROM
+                                "User" u
+                            JOIN "Metadata" m ON m."userLeetcodeUsername" = u."leetcodeUsername"
+                            WHERE
+                                u."leetcodeUsername" = ?
+                                AND
+                                m."leaderboardId" = ?
+                        """;
+
+        try (PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, userLeetcodeUsername);
+            stmt.setObject(2, UUID.fromString(leaderboardId));
+            try (ResultSet rs = stmt.executeQuery()) {
+                if (rs.next()) {
+                    var id = rs.getString("id");
+                    var discordId = rs.getString("discordId");
+                    var discordName = rs.getString("discordName");
+                    var leetcodeUsername = rs.getString("leetcodeUsername");
+                    var nickname = rs.getString("nickname");
+                    var admin = rs.getBoolean("admin");
+                    var profileUrl = rs.getString("profileUrl");
+                    var totalScore = rs.getInt("totalScore");
+
+                    var tags = userTagRepository.findTagsByUserId(id);
+
+                    var userWithScore = new UserWithScore(id, discordId, discordName, leetcodeUsername, nickname, admin, profileUrl, totalScore, tags);
+                    return userWithScore;
+                }
+            }
+        } catch (SQLException e) {
+            throw new RuntimeException("Failed to get user with score by leetcode username", e);
         }
 
         return null;
